Add files via upload
Financial and Tax Due Diligence is an essential process to determine sell side and buyside quality of earnings, that kicks off after the signing of LOI in an M&A transaction. Due diligence entails a lot of data, asked from the business owners and their core financial team. Due Diligence team conducts official visits to the Business owner, and usually takes a checklist of questionnaire with them. Although there are many non business related expenditures too that are discussed, to arrive at a business valuation, this list of imperative questions reduces back and forth done between the Diligence team and the business owners. It helps to reduce time spent on compiling and organizing essential data, so that much of the allotted time can be spent on analysis of the business earnings. Analyzing quality of earnings is beneficial to both the sell side as well as buy side businesses.

Benefit to Sell Side: 
The business owner will be aware of both positives and negatives of the business, information components that may not have been looked into before, keeping the sell side prepared for a conversation when there is a potential buyer, ahead of the time where pain points might be.

Benefit to Buy side:
PE firms can unearth information to mitigate legal risks, identify profitable deals, make informed decisions, discover synergies and promote transperency and collaboration.